Merge translations of Bluetooth confirmation text
This is a manual merge of the translated strings from the master branch
of the 'bluetooth_pairing_shares_phonebook' string. The fix for this bug
requires these translations because the fix was to slightly reword the
confirmation text in the pairing dialog and stop substituting in the
device name.

Bug: 62672248
Test: manual
Change-Id: Ia386cc2d98de086ed545e4fe6d41be20e9c24dc4
(cherry picked from commit 7d505aefa59f87e2fb7cce57d4d5fad41a1b51aa)