Merge translations of Bluetooth confirmation text
This is a manual merge of the translated strings from the master branch
of the 'bluetooth_pairing_shares_phonebook' string. The fix for this bug
requires these translations because the fix was to slightly reword the
confirmation text in the pairing dialog and stop substituting in the
device name.

Bug: 62672248
Test: manual
Change-Id: I829baada46cdb2314021d8772feb8811072e41d8
Merged-In: I2d8c57680d1d98b74193654e9fe77016304f6ac5