Merge translations of Bluetooth confirmation text
This is a manual merge of the translated strings from the master branch
of the 'bluetooth_pairing_shares_phonebook' string. The fix for this bug
requires these translations because the fix was to slightly reword the
confirmation text in the pairing dialog and stop substituting in the
device name.

Bug: 62672248
Test: manual
Change-Id: I049138002f07e78442ddebfa746d6128b5689947
(cherry picked from commit a0b8f3fe3aed0c66daa4bb9422a28c56f2dea6b7)