Add a default preference controller for developer options.
- some of the preferences in the developer options page do not have a
preference controller associated with them, because they are simple
preference that do not require any update on the preference state at
all. However, they still need to be enabled/disabled when the master
developer options switch is turned on/off. Adding a default controller
to be used for these preferences.

Bug: 73955632
Test: manual
Change-Id: Ifb757dee2e5957d266ebb6ad5421d8ff9adbb142