Don't allow a zero-length device name to work.
Technically, if a device has a zero-length device name, the bug will
resurface. The EditText validator doesn't trigger on opening -- only
when the text is edited. A zero-length device name is flagged as being
invalid, but that fails if the text box starts empty.

By pre-filling it with the previous device name, we can ensure that, as
long as a zero-length device name is never set, it can never be set.

Change-Id: I0d28aaae09f99b7d697b753835ba39c0c06644a1
Fixes: 73127912
Test: Robotest
(cherry picked from commit b2165509445c0c63b1813132d8b92b119bd44d57)