Update add network dialog to not make networks hidden by default
This CL updates the add network dialog to use the provided value
for the hidden state (from androids perspective) of the network. The
default selected value will always be not hidden and if the user
selects to mark it as hidden a warning will appear to inform the
user of the privacy implications.

Test: Robotests
Bug: 29003359
Change-Id: Ifd4511e4d0af6eb01f4e951c22a91fc560393e8d