Make it possible to have open tether network
Previously people could have open tether networks but a restrction
was placed on this sometime in the past. However there is no migration
plan in place for people with open tether networks which will cause
problems when the new requirements are enforced. This CL makes it
so tether networks can be open once more.

Test: robotests
Bug: 65784990
Change-Id: I6e350dd53b9b9f987dd5a4cc317ba18d7f50df79