Update settings to reference correct versionCode
This settings change is required for a framework change that ensures
that apps built for pre-P that rely on reflection to access
ApplicationInfo#versionCode don't crash. The move to long version
code introduces a new field and all modifications of the field are
wrapped in a method that ensures both the new and old fields are set
appropriately.

Bug: 74393568
Test: manual - builds and broken app runs
Change-Id: Idfa5f85d3f91583098ebee88f0e8caecaacff9b4