Clean Sms raw table when resetting network.
There has been issues when Sms raw table somehow stores orphan
fragments. They lead to garbled message when new fragments come
in and combied with those stale ones. While we work on proper
fixes, one of the options we offer to the user if it happens again,
is to reset all network settings which will clean up this table.

Bug: 77910620
Test: manual test
Change-Id: Ib80c395fb9a1501f35716fdca43df3e5c87893c7