Fix RTL alignmnet of "eSIMs" in "Erase all data" page.
"eSIMs" is not a RTL string, so it will align to left even we change to
RTL language like Urdu. We must set the textDirection of TextView to
"locale" to force it follow locale's direction. We don't use
textAlignment to set string direction because it contains list bullets,
and list bullets will still align to left.

Change-Id: I0eea2bed9f324bb071e413673227839b830921cf
Fixes: 77946796
Test: visual