Layout changes for new animation assets
The animation videos are changing to have a white background, so we want to
match that in the background color to avoid horizontal bands of the current
grey background color on the edges of the new videos. We're also adding a
divider between the video and the content beneath it.

This CL also includes the new illustration video for the launcher gesture.

Fixes: 77861241
Fixes: 77474484
Test: manual (visit all entries in Settings->System->Gestures, as well as
Settings->Display->Adaptive brightness)

Change-Id: I976782645c9431d40e005960b80353a7da2b89b8