Avoids systemui color tint flicker because no system ui flags set
Systemui nav buttons set the insensity from launcher and -1 window but
fallback home does not. This causes services to pass the wrong flags to
system ui to apply the incorrect tint that would later be fixed by
launcher and sometimes lead to color flicker.

Fixes: 79776583
Test: use bright color wallpaper (nav dark icons) and reboot phone,
unlock with pin/pass/finger and view nav buttons

Change-Id: I86906cd1b5af5d078499a1392c260aa186b2eee8