Remove references to SettingsBase styles
We've moved the values for several preference-related styles used in
Settings from SettingsLib into the support library. This CL removes the
references to the SettingsLib versions from Settings so they can be
removed from SettingsLib.

Also delete an unused layout file that was moved to SettingsLib a long
time ago (preference_dropdown_material_settings.xml).

Bug: 64689839
Test: visual (styles for things like icon spacing, dividers, etc. should
be the same)

Change-Id: Ib7a745218086269c2e861b18df68bd835929da19