Fix issue that read-only apns can be edited
When entering to ApnEditor with ApnSettings#onPreferenceTreeClick(),
subId is not passed.
So invalid subId is used and it gets wrong carrier config values.

To resolve this issue, remove unnecessary ApnSettings click event.
The only click event for launching ApnEditor is ApnPreference click
event.

Test: manual - Check that the read-only APN can't be edited when tap
the edge of the APN settings screen
Bug: 117302653
Change-Id: I0a61b409539d6e4f6aa9860235e61c43c76640fa