Launch work profile activity to add Autofill service
ag/4666330 allowed choosing separate Autofill services for the main and work profile. However, the 'Add service' button for work profile always resolved action_view intent as the main user even when clicking the work profile preference. Now the intent is resolved to the right profile depending on the preference clicked.

BUG: 112610177
Test: Verified manually
Change-Id: I870346d7b5618e7c230edebd59ac7f52f8495d47