Add a feature flag for new version of Network and internet settings page
In addition to adding the flag, this also adds a new xml layout that is
the same as the existing network and internet page, but will begin to be
changed in subsequent CLs. There is also a change in code to use this
new layout when the flag is turned on.

One other change is to make the list of displayed feature flags in
developer options be in sorted order instead of whatever semirandom
order they come out of the map's key set.

Bug: 116349402
Test: Manual (settings_network_and_internet_v2 flag should should up in
Settings -> System -> Developer options -> Feature flags, and those
flags should be in sorted order)

Change-Id: I5520d286efb6bec50ba4ca99f64aeca94f01d533