Fix app link open mode UI
The app link open mode UI misrepresented some states; notably, it
was representing the 'ask' state as the 'always use this app' state.
This had the effect of implicitly pushing the app into 'always'
accidentally, as a consequence of setting the UI to that initial
state, then re-evaluating on exit against the true state.

Bug: 32810168
Test: run cts -m CtsOsHostTestCases
Test: manual
Change-Id: Ifb9d7ad0140fae5fa75a4b78b01301b196a48d10