Add wifi connection info to the multi-network header
The Network & internet page will have a dynamic header at the top when
users have more than one mobile subscription, showing information about
connectivity.

This CL adds a preference to this header when there is a wifi connection,
showing the same information as on the wifi-page (connection strength,
speed rating if available, etc.).

Bug: 116349402
Test: make RunSettingsRoboTests
Change-Id: Ia80d6e236a4996b501372ac4cd8e46ba6c5f8841