Polish some behaviors of the battery saver schedule screen
This CL removes the line that was showing between the radio
buttons and the seek bar. Additionally it makes it so the seekbar
updates continuously as a user moves it and fixes a visual bug
where it would show 5% regardless of what the user had set it to
when exiting then re-entering the screen.

Test: visual
Change-Id: Icdb9cafe2bdda03d94230d1c8538f4c322b050e9
Fixes: 121388626
Fixes: 121211075