Set summary for SIM entries in multi-SIM header
On the Network & internet page, when you have multiple active SIMs we
add a header with an entry for each one. This CL sets the summary for
those entries to be a 1 or 2 line indication of what services (calls,
SMS, and data) that SIM is used for by default, or just an indication
that the SIM is available. And for the SIM used by default for data, we
include status information about whether it's connected or not.

Bug: 116349402
Test: make RunSettingsRoboTests
Change-Id: I46f37433f753724bbde8876f0d89c3155b2c3812