Fix Slices in Panel no-op bug
Slices in panels stopped changing the underlying settings.

It is not always reproducible, but the error happend when PendingIntents
started failing to fire, and thus not connecting with the Broadcast
Receiver in Settings dedicated to handling Slice changes.

This fix always highlighted that we shouldn't be using TaskAffinity,
because there is no guarantee 3p apps don't spoof it.

In order to fix both the PendingIntent issue, and to de-dupe activities
from Settings vs Settings Panels: we use 'documentLaunchMode=always' instead.

Fixes: 129006165
Test: manual
Test: atest SettingsPanelTest
Change-Id: If94dd719a2e33c85c04a416868f565f719941234