Add a string for prompting which SIM to use for SMS
In some multi-SIM cases the telephony stack will fire an intent that
Settings listens for, specifying that Settings should present a dialog
to the user asking them which SIM card they want to use for various
services such as Data, Calls, and SMS. The string we were using for SMS
was a general "Select SIM card". This CL adds a more specific one that
should make more sense to the user.

Bug: 129021763
Test: make Settings
Change-Id: I8c422a7e6de3ffd1255bc7b4124d9a1be6adf633