WifiCallingSettingsForSubTest: Solve casting problem in setup
Solves casting problem in WifiCallingSettingsForSubTest that causes all
tests in this file to fail in the setup-function.

The wfc mode buttons has changed from being a ListPreference to be a
ListWithEntrySummaryPreference. The test cases have now been updated
accordingly.

Problem was introduced by the following two patches that conflict with
each other. The test cases in WifiCallingSettingsForSubTest works
correctly when each patch is tested in isolation, but failed when they
were both merged:
73ffcf49990ab3134b139607c0225f6ae1646c18
8498436a9613c806d4e1085074aea22f1a78675a

Bug: 129545431
Test: make RunSettingsRoboTests \
ROBOTEST_FILTER=WifiCallingSettingsForSubTest
Change-Id: Ib6988e514bbad2023ba6cfc2e1d8734b68cabc31