Tint icon colors of storage screen same as other screen
We were using hard-coded colors to tint the icons on storage screen
for some reasons.  Change it to the same way we tint other icons,
as there's no reason to keep it special, and also per partner's
request to keep it consistent with other screens.

Also update the ic_sim_sd icon from png to vectors.

Test: Manual/Visual verification
BUG: 113977374
Change-Id: Ia0210635e936b6798aaf9e48dd8815a197396065