Make RadioInfo settings menu work for mSIM
Just add a spinner to the top of the screen which lets you select the
phoneIndex. Upating this value updates the mTelephonyManager
mImsManager, and mPhoneStateListener to use the selected phoneIndex and
the subscription associated with it.

Also adds fields for current subId and default data sim subId.

Due to b/117555407, the same PhoneStateListener object can not be used
to register on different subIds, so we workaround this by recreating
the PhoneStateListener when reregistering.

Test: manual test (on single sim)
Fixes: 128033739
Change-Id: Id075e3bffe20523d8ceeb8f2e4320f995abdab55
Merged-In: I9d6631da628351511e285afeb5b6d4331aaab7e5