Create TintDrawable in Settings.
This allows us to specify a drawable resource in xml that is simply
another drawable with a different tint. This is necessary to overlay the
settings wifi icons since the tint is resolved using a theme attribute.
This change also replaces a few icons other duplicate icons to
demonstrate usage.

Fixes: 123363812
Test: make; use class in XML and run app
Change-Id: Id80f318e96dab2e1ad9db8430e8371ac6497b753