Update listDivider for PreferenceFragmentCompat
The support library uses preference_list_divider_material for its list
divider and its color is a hardcoded value. That's why the divider color
can't be applied to the PreferenceFragmentCompat in Settings when
updating the divider color in the device defaults theme level. To make
it so we set its listDivider to list_divider_material to facilitate updating
the divider color.

Bug: 129943610
Test: visual
Change-Id: Ie992b925fae0effd4e00682a911ad004ac3fc033