App Bubble control page should always be available, even if off globally
There is API to open up the app-specific bubble page, therefor, we should
always show the toggle. The previous work to move bubbles to developer
settings broke this.

This CL adds a param to the controller to note whether it's the app
specific page or not (since controller is also used for channel settings)
to enable this behaviour.

Added a test for this situation.

Test: make -j40 RunSettingsRoboTests ROBOTEST_FILTER=Bubble
Bug: 131846917
Change-Id: I72b1f4b5b033d1b24281061cfa6bed4d734dfcb1