Fix flaky test error
Before this CL, we didn't consider the case of two
preferences timestamp are the same.

If 2 preferences timestamp are the same, the second preference
will first out. it will cause test case will fail some times.

In this CL, if two BluetoothDevicePreference's timestamp are the same,
The first BluetoothDevicePreference is first out.

Bug: 138547532
Test: make -j42 RunSettingsRoboTests
Change-Id: I7366275e8edf615c582481a570ea0c25b8d37a66