[Manual Network Selection] UX improvement
Original design abort a network scan when screen off and restart scan
when screen on.

Under the situation of network scan time longer than screen off time,
user may not get any result but keeping in scanning screen.

This CL improve this situation by avoiding stop and rescan for the case
user didn't leave this UI and previous scan is in progressing.

Bug: 140448617
Test: Manual testing leave and back to UI, and screen off-on.
Change-Id: I74ffa574dac932ca7fcd76411cad4ec0cc88f25f