Optimize NetworkScan for only User-Enabled RANs
Network scans comprise multiple different access
technologies, and scanning each of those techs can
be very time consuming. Worse, depending on a user's
settings, the device may only be allowed to connect
to 2G, 2G+3G, or 2G+3G+4G networks. If a scan finds
network types that a user can't connect to, it also
creates a poor user experience if a user tries to
select those network types.

This CL restricts the scans to only find and return
results that a user can select based on the other
user settings.

Bug: 128572818
Test: manual (crosshatch) - set allowed networks to
 2G+3G and verify that LTE networks are not returned.
 Set allowed networks to 2G+3G+4G and verify that LTE
 networks are returned.

Change-Id: Ic855100723b8604c1e872457a70fd8f25d225278