Added a warning dialog when installing a CA certificate
This is part of the changes to improve the UX and language for installing certificates.
Previously, the different types of certificate used the same installation flow.
Due to concerns around users installing CA certificates without understanding the conseqences,
this CL introduces a new warning dialog when a CA certificate is installed from settings.

Bug: 139173976
Test: Atest com.android.settings.security
      manual testing from Settings by selecting the certificate type
	preference and ensuring the installation flow still worked as expected.

Screenshot of the screen: https://hsv.googleplex.com/5046848484016128

Change-Id: If95bffd1e68f14734fb20e8cc4b60eeb1c372358