Move the app to /system_ext
As mentioned in go/partitions-in-r, we plan to forbid hidden API in
product partition.
So this app should be moved to /system_ext because it uses hidden API.
It is not permanent, so they can be moved back to /product due to
needs if there isn't hidden usage anymore.

Bug: 137908189
Test: m
Test: check whether device boots successfully
Exempt-From-Owner-Approval: cherry-pick from internal
Change-Id: Icd2f7fc7021e84fad9980d58688dc640ca4245f8
Merged-In: Icd2f7fc7021e84fad9980d58688dc640ca4245f8