Allow any app to launch AppManageExternalStorageActivity
This activity is started by launching an intent with the action
android.settings.MANAGE_APP_ALL_FILES_ACCESS_PERMISSION and the package
name.

Previously, we required launching package to be granted the permission
MANAGE_EXTERNAL_STORAGE. This is not WAI because we want any app that
requests the permission to be able to launch the activity.
Because the activity allows apps that have requested the permission to
toggle the special app access anyway, this CL drops the requirement
mentioned above and now any app can launch the activity.

Test: manual - a normal app that doesn't request the permission can
launch the intent, and it can't toggle the special app access.
Test: manual - a normal app that requests the permission can launch the
intent, and it can toggle the special app access.
Fixes: 147906545

Change-Id: I605a5b03f9be1b4b11d94e121a18bbd316404590