Set ClearDefaultsPreference as unselectable for talkback
ClearDefaultsPreference is a customized preference which contains a
TextView and a Button. In reality, the button is the only interaction
way to users. There are no more interaction when user click the whole
preference view.

Mark the ClearDefaultsPreference as unselectable can fix the talkback
issue for user clicking the preference view out of button.

In the meantime, the button of ClearDefaultsPreference is not yet to
apply the button style of Settings. Also fix it with ActionPrimaryButton
style of Settings.

Fixes: 146471192
Test: visual and robotests
Change-Id: Ife215d0901bd9fc3055495b0495ba175ed0dfa26