Add developer option to disable automatic adb grant revocation
Android 10 introduced a security feature to automatically revoke
adb authorizations for systems that have not reconnected to the device
within 7 days. While this is helpful for consumers that enable adb for
a one time task and mistakenly select the 'always allow' option,
feedback has indicated having a developer option to disable this feature
would be beneficial.

Bug: 119510647
Test: make RunSettingsRoboTests ROBOTEST_FILTER=AdbAuthorizationTimeoutPreferenceControllerTest
Change-Id: I7eb123e8c69956aa02bb679784ac79650baf5dcb