Enable Night Light preference searchable
When the searchable attribute of preference is false, we do not collect
it to Settings Search, therefore, we can not get its sitemap relation
from index data.

Since we do not index the NightDisplayPreference, it caused the sitemap
relation loss. The relevant night light preferences can not map to
the correct parent page, Display page.

In order to fix the breadcrumb and grouping problem, remove the
searchable="false" of NightDisplayPreference to enable the indexing
again.

Bug: 151206664
Test: visual
Change-Id: Ia4f82b29cfe4f78ae55b20dad0356ac56b735307