Only request focus if change was self-triggered
For a11y we want to accomplish two things:
1. Keep focus on the button if the button was just pressed. This
requires buttons in settings to be focusableInTouchMode and then request
focus when the buttons are updated.
2. When the setting is changed outside of the settings page (ie: from
Quick Settings), we don't want to re-announce on talkback the button that is
updated on the Settings page. Therefore, to maintain #1, we add a
boolean to check whether or not the updateButton change was triggered by
a click to the button preference or not. Only request focus if the
button was triggered by the settings button (and not when triggered by
other things like QS).

Test: manual
Fixes: 147992252
Change-Id: I505de0585bac38a2f59062743d9d6b75861c3d32