Update display logic for no shared data.
When there is no shared data, instead of showing an empty list view,
the developer option's summary is updated.

Additionally, for the two shared data screens, added the back button to
the action bar so that it matches the other screens in the Settings app.

Also, implemented PreferenceControllerMixin for the
SharedDataPreferenceController so that the dev option shows up in
settings search.

Bug: 153079829
Test: make RunSettingsRoboTests ROBOTEST_FILTER=SharedDataPreferenceControllerTest
Test: manual (visual)
Change-Id: I6ec5c35fc262dc27a383af8e2593be5b349ee85f