Fix Auto click (dwell timing) does not speech out meaningful delay time in enlarge or scroll by volume key
Root cause: does not speech out the auto click time when adjust the delay time
Solution: Add android:accessibilityLiveRegion in the title of customized scroll view. When the delay time is changed, the text of title will changed, too. It can use the attribute to make accessibility services announce changes to this view.

Bug: 154295215
Test: local test
Change-Id: I1eceae310251c586bbb74185828da0ef49bfae99