Add icon to current screen saver preference
The screen saver setting is coming from different apps to modify
their options. This will confuse people that they were modifying
something in Settings rather than in other apps. To improve that,
we added an icon of active dream and put it on the left side of
current screen saver setting, which is to let people clearer to
know that they were modifying the options in other apps.

Bug: 129994275
Test: robotests
Change-Id: I014c9de745044c2bd39d22782b6c36f8605d8f56