Remove index for Take call on of Sound page
"Take call on" is only visible when there are multiple media devices
connected and phone call is incoming or answering. In general, this
preference is hard to see for users.

In reality, users usually change the phone call output on Phone app,
it's a rare case that user will go to Settings to search "Take call on"
to change the phone call output when they are answering call.

Intend to remove its index from Settings Search first since it's almost
invisible on Sound page. Users will confuse that why Sound page doesn't
show the corresponding preference when they found the "Take call on" and
tapped on it from Settings Search.

Fixes: 140392320
Test: visual and robotests
Change-Id: I87ec4fda2a3d8625c8f09beeffe348a302187d6d