Haptic effect on swiping the notification item
Haptic once the swiping on the notification item is going to snap in either directions. The snap-in scenario is about the notification item when there is a "snap back point" i.e. if swiping the item back till a certain point it just snaps back to initial state but once it goes past a certain location it snaps into the new location.

- screenshot, https://screenshot.googleplex.com/6A8Gxs7yRwqAVk2

Bug: 175364588
Test: manual
Change-Id: I7e2ed19bfb7f863502e10233e3e23ee5d434b3b4