Assign StaticIpConfiguration builder value to global variable
[Root Cause]
Due to call by reference for Object, we pass the StaticIpConfiguration
to method. But in method we new the staticIpConfiguration builder
and re-assign builder to parameter, it will release when the method is ended.
The value in Builder will not be set to global variable.

[Solution]
Assign the value in StaticIpConfiguration builder to global variable
directly.

Bug: 186489618
Test: manual test
      make RunSettingsRoboTests ROBOTEST_FILTER=WifiConfigController2Test
Change-Id: I1a6aee31e6977be42792440c3667db2557a20c4e