[GAR] fix spoke a percentage number is not the same as displayed content
in the Battery Saver and Battery Share, we have customized rule to map
the seekbar progresss to another displayed percentagge value, which
cause the a11y will speak the incorrect state, since the state is
referenced the original progress value. we add a method to override it
into our cusromized value.

Bug: 187780942
Bug: 190958777
Test: make SettingsGoogleRoboTests
Change-Id: Ie630ac03e66c2f8da1df00d6d2616b2e6979aa3e