ManageablePreferenceTest created for VpnSettings
Creating unit tests for VPN settings to increase the reliability of that
settings component.

Note that because S build is locked down but the T sdk has not been
started, the test suite was seeing the following error:

java.lang.IllegalArgumentException: Error: Failed to parse APK file: /data/local/tmp/SettingsUnitTests.apk: Requires development platform T but this is a release platform

In order to resolve this error, specifically set the SDK as android
S in the manifest. When the SDK transitions to android T, this may need
to be updated.

Test: atest -c tests/unit/src/com/android/settings/vpn2/ManageablePreferenceTest.java
Bug: 187245804
Change-Id: Ifa1450c23b1ade46809541252e22950b76b4faa5