Fix multiple biometrics enrollment flow
Suppose both face auth and fingerprint auth exist in a device, the setup
flow will skip fingerprint enrollment when a user skips to register
face data on the face enrollment education page. Updated the skip button
method and make the flow move on to fingerprint enrollment if a use
doesn't want to register face data.

Fix: 186084024
Test: manual verified
1) Reset device and start from SUW
2) Press skip button on the face unlock education page
3) The next flow should be fingerprint enrollment intro page

Change-Id: I037ae68f39f5fdbf62ea95b998621c7f7df31b43