Correct the choose screen lock page
The choose screen lock page didn't apply the right theme during the
enterprise flow. That is because an intent extra of setup wizard wasn't
properly passed to the next page. In this change we made sure the intent
extra is able to properly pass to the next page.

We also removed the wrong theme for setup choose a screen lock page
since the wrong theme made this page broken. The setup choose a screen
lock is a sub setting page that wasn't implemented by using SUW library,
so it should not use GlifLayout theme. With this update, the choose a
screen lock page still keeps the background color of Settings, that is
different from the one of setup flow.

Bug: 190499041
Test: manual test
Change-Id: I9dceee6a398c7e7b487dd8e4046f71f76cc50e36