Revert to Default Settings
Add functionality to the revert default settings button. This currently
writes a null value to the ENABLE_TARE constant. In the future this will
also make the AlarmManager and JobScheduler constants null so that the
default values are used instead of custom config values.

Bug: 195543827

Test: Click on revert to default settings button, check that the value
for ENABLE_TARE is null by running "adb shell settings get global
enable_tare" in the command line

Change-Id: I558e8f1121021eca3224e4bc9b26bbe0dd2467c7