Start Safety & regulatory manual with new task
Based on the large screen design, Safety & regulatory manual should be
launched as a new task since it doesn’t fulfill the Settings UI design.
And it is not suitable to customize the background and font style with
Settings since the manual is provided by the legal team, we should keep
the content and style aligning with the server side.

Fixes: 197702066
Test: visual
Change-Id: Icaba237961cc774ed1bdca50d9095dc5409039e8