Fix abnormal behavior on avatar account page
Prior to this cl, if user opens settings app
in single-pane first and navigates to
the avatar account page, then rotate the device,
user observed the account page was still shown
with full screen.

Because we didn't register correct split rule,
it causes the abormal behavior on two-pane mode.
In order to register correct rule,  we also need
to assign correct component name while opening the
account page.

Bug: 207609699
Test: Rebuilt apk and verify the behavior
Change-Id: I31def684c033c1d0c20870284826c6713a31b43d