Fine tune Caption size and style preview height for tablet
Root cause: We introduce the material design illustration in Android S. The most of illustration preview height is 300dp, but the SubTitleView preview keep 200dp in the phone and 150dp in tablet. The value is defined in 2017.
Solution: Align material design illustration height which looks a lot closer to an actual video, without pushing too much content below the screen.

Bug: 227264630
Test: Manual testing
Change-Id: I1a9040f50dfb6fb95c1cadbe4e7e4d24ff8ab058