Fix Wi-Fi "Network usage" and "Privacy" drop down
Currently, these are implemented with DropDownPreference, which keeps
disappearing after opening. This is because this page is special, many
other preferences on this page auto refreshed every some seconds,
causing the container RecyclerView to re-render (and a scrollbar is
displayed when re-render happens). This re-renders cause the drop downs
to auto dismiss.

Use ListPreference instead can solve this issue, because
DropDownPreference is extended from ListPreference, this is simple
change.

Fix: 209799515
Test: manual
Change-Id: Ib098922c39d9d6c56d645f12f5884489ea6688be