Improve Talkback for user select dialog
Before this fix,
Both the list item and button are focusable by Talkback, and the list
item one is ok ("Personal, one of two."), the button one is read as
"Unlabeled".

After this fix,
Only the button is focusable by Talkback, read as "Personal, in list, 2
items."

Also test to make sure Switch Access not break.

Bug: 174626616
Test: manual test when Talkback / Switch Access enabled
Change-Id: I72765c1bcbb75e544d5829b21c3e2baf4355be1f