speech: Add on-device speech recognition settings entry
Currently under `Settings > System > Languages & input > Speech` there
is a `Voice input` entry with a cog which starts the SSBG model manager
settings activity. Under the same preference group another entry
labeled `On-device recognition` is added which would open a similar
model manager settings activity, but for the on-device recognizer. That
settings activity is yet to be implemented.

The new entry should appear only under the following conditions:
 - An on-device speech recognition service must be available in the
system.
 - A speech recognition service with a proper settings activity
meta-data must exist in the same package as the default on-device
speech recognition service.

Bug: 235457391
Test: Manual, already existing robotests
Change-Id: I17208c8725500ccb3dd2fa51a12b003d32073c4e