Add field to control margin in FaceEnrollIntro
Problem -> The require scroll mixin seems to scroll an entire device
length when a button click occurs. If you have page content
of 2x device_length + 10 pixels, a user will have to press the more
button twice to scroll this length. With the illustration_margin_bottom
field, devices can tune this parameter to allow the user to press
the more button once.

Fixes: 237474002
Test: Manual.
Change-Id: I9de2d74608260fe835870211a07b96249a8598f8