[MEP] Hide the preferred SIM dialog when the user has replaced the SIM
Since there is the race condition and it causes UI hides the preferred
SIM dialog. Therefore, to hide the preferred SIM dialog under the
specific condition which the user has replaced the SIM during the
SIM switching.

Bug: 238061853
Test: Manually testing. Device has the psim+esim and the esim's mobile
data on. The tester disables the esim and then UI shows the preferred
SIM dialog.

Change-Id: I01e7d60170c5053730fd3113abd914fb5c0d11c9