Prevents retrying fingerprint after lockout reset if user exits FingerprintSettings
Fingerprint auth is active on the enrolled fingerprints page to indicate
to users which fingers are enrolled. Tapping 5x with unenrolled finger
on this page sends FINGERPRINT_ERROR_LOCKOUT, which posts a
mFingerprintLockoutReset runnable to retry fingerprint auth after
LOCKOUT_DURATION. If a user turns off their device and unlocks via
pin/pattern/password, this resets lockout, triggering the runnable and
causing fingerprint auth to be retried upon device entry. This change
prevents this by removing the runnable callback in
FingerprintSettingsFragment#onPause.

Test: Navigate to enrolled fingerprints page (Settings > Security > Face & Fingerprint Unlock > Fingerprint Unlock). Tap 5x with unenrolled finger, then turn off screen and unlock via pin/pattern/password. Ensure indicator does not appear upon device entry.
Fixes: 261366312
Change-Id: I4190669286461e9ce884685a9929313da953476d