Removed isWorkProfile_withSystemUser_returnFalse test
Only the UM.isManagedProfile property should be used to determine
whether a user is a managed prfile.
There is no need to perform any additional check on top of
UM.isManagedProfile to make sure the user is a managed profile.

After removing the additional check as part of ag/21512484
DataBaseUtilsTest.isWorkProfile_withSystemUser_returnFalse test
is redundant.

Test: atest DataBaseUtilsTest -c
Bug: 270283369
Change-Id: Ic4fd7eef2ac24786e2a59db7dabc9daf4a7ce028