[Settings] Fix the layout index
The index of language label's layout is 0 and index of num layout is 1, so the select tp speak will read the language label first.
The root cause of the incorrect index is, the language label's layout starts with the parent layout instead of the num layout in the hierarchy view. Adjust the language label's layout and add these views one by one in the parent view.

Also fix the layout's jumping when user switches the page to "Remove language"

Bug: 277173513
Test: local test
Change-Id: I0d5879a07183713fa50b38f6321a15dd32582518