work challenge: allow backgroud biometric auth
We observed a race condition that when work challenge is launched from
notification, the biometric prompt quickly dismisses itself because
it detects the launcher is now the foreground app. This change attempts
to workaround the issue by enabling the setAllowBackgroundAuthentication
optin in BiometricPrompt so it no longer dismisses itself even if the
foreground app is different.

Bug: 279766640
Test: manual
Change-Id: I453b7d603c6eb65f329afb38d8a190e21a7e4c01