Allow BAL from PrintServiceSettingsFragment
This fixes the info button on the printer in the printer list (the call
is from the system (uid=1000) and needs to allow BAL to start the
activity).

This only opts in sending a single pending intent to the previous (pre U) behavior which allows BAL. Therefore the risk of the change is minimal. This was reported by an external partner.

Bug: 277761614
Test: manual test & presubmit
(cherry picked from https://googleplex-android-review.googlesource.com/q/commit:fdb87fe235887460081ed396cf6ab902f1971b5d)
Merged-In: I35715e948b9166c42372b7d40a30a7102018a148
Change-Id: I35715e948b9166c42372b7d40a30a7102018a148