Set a11y focus for selected Settings item
When a Settings preference item is selected via a settings panel or
settings search, the system takes user to its corresponding subsettings
activity. There, in the past, the selected preference was highlighted
but the a11y focus was set on the back button. This code sets the a11y
focus on the selected preference.

Bug: b/285512439
Test: atest HighlightablePreferenceGroupAdapterTest
Change-Id: I7edc697e2e3870b1822279c115230add14e82df0