chore(text reading settings): increase preview block left/right padding when landscape mode for easier scrolling
When the device is in landscape mode, In the text reading preference page, users have difficulty scrolling the RecyclerView because the preview block is almost full-screen. To scroll the preference list, users typically try to pan the preview block, but the block consumes the scroll gesture, preventing the list from scrolling. Users can successfully scroll the list content by panning outside the preview block. As a quick workaround, we increase the left/right padding of the preview block for landscape mode to make it easier for users to pan outside the block and scroll the preference list.

Bug: 301683714
Flag: NA
Test: manually - attach screenshot in bug
Change-Id: I1dba6e4d83ee1cd2f30163de05dfd7b42f829e42