Settings changes for changing default for zen policy.
This change sets a rule's ZenPolicy to the device default right before entering the customization screen, rather than setting it to a copy of the global DND policy. Further, because all policies are now fully specified, when the MODES_API flag is on, skip the radio button screen for selecting between a default and a custom setting and jump straight to the page for setting custom settings.

Bug: 312739308
Test: manual -- entered settings for rules created before and after V, modified settings, verified they took

Change-Id: I1e3a777b60a8fcf06732879f823b033826a52a43