Restart or finish HomepageActivity when it's launched unexpectedly
Settings homepage should be started with the flag FLAG_ACTIVITY_NEW_TASK
to ensure it's the first activity in a task. Otherwise, there will be UI
issues.

When homepage is started with FLAG_ACTIVITY_NEW_TASK but not the first
activity, it means another homepage activity has been started, so
finishing the current invocation will bring up the previous activity in
the task.

On the other hand, when homapage is not started with the flag and not
the first activty, it indicates that the invocation is from another
app. Restarting the homepage activity with FLAG_ACTIVITY_NEW_TASK will
keep the UI behavior consistent.

Fix: 297857732
Fix: 309045575
Test: robotest, manual
  1. Start Settings from QuickSettings, go to any subpage, navigate to
     home, click Settings icon on the launcher, and the previously
     opened page should be kept.
  2. On a LS device, start Settings from another app without
     FLAG_ACTIVITY_NEW_TASK, Settings should be started in dual-pane
     mode.
Change-Id: If1f31e26dc37f681bd97c185cbeac2de06bdd48f