[Audiosharing] Update the volume control registration flow.
Based on the framework API redesign, we will start/stop the volume control callback registration when user enter/leave the volume control page.
We will temporarily keep a map of the device to volume map and use the
volume value to set the initial state of the volume progress bar if it is valid.
If the volume value is null or invalid, we will read the music stream
volume from the audio manager to set the initial state.

Test: manual
Bug: 305620450
Change-Id: Iee3fba39af388782ac3498cb1610b248ced22c3c