Add biometric enrollment support for private profile
During a new lock setup for profile whose credential is shareable with
its parent first the user is authenticated with device lock after which
an activity having options of Pin/Pattern/Password with fingerprint and
face combinations is shown.
On choosing any option which has combination to set LSFK and biometric
it is expected that after setting LSKF the Biometric enroll activity is
started but currently this does not work as expected as the
ChooseLockGeneric activity is finished after adding LSKF and it does not
start the biometric enrollment for the profile.
The issue also exists with non-profile users using this workflow through
SET_NEW_PASSWORD intent and if already have LSKF assigned.

This change adds a new boolean which takes care to not finish the
activity till the Biometric enrollment is started.

Below conditions are taken care with this change
- For new lock setup when device lock already exists then after
  authentication of current device lock make sure the activity is not
  finished untill the biometrics enrollment activity is started.
- On choosing continue without fingerprint or face option the biometrics
  enrollment is not started

screen recordings uploaded to buganizer - b/316109077

Bug: 316109077
Test: Manual
Change-Id: Ifcbaa7d89195d87d432fc848092f2301752c3c22