Use SetScreenLockDialogActivity to power screen lock setup prompt
With this change, we move to using SetScreenLockDialogActivity to enable
the prompt to set up the screen lock. This prompt is shown whenever
someone taps the private space settings page option to either
setup/unlock the space. We will also be using the same dialog prompt in
other places in the framework
(UserManagerService.requestQuietModeEnabled) as well.

Test: atest PrivateSpaceAuthenticationActivityTest
Bug: 316129700
Flag: android.multiuser.show_set_screen_lock_dialog
Change-Id: Ib9fe009d3b36e0eea242fbf894e616b1efcb9d6b