Touchpad: remove icons
We don't have an icon for the tap dragging setting, and settings screens
should either have icons for everything or nothing, so we decided to
remove all the icons. For all except ic_trackpad_touch_gestures_inverse,
they're only used in this one location, so the drawables can be removed
as well.

Screenshot: go/b-324058706-icon-removal-screenshot

Bug: 324058706
Test: with a touchpad connected, check that the System > Touchpad screen
      has no icons on it.
Change-Id: I0c0a76c757f568cd20799c53a96cfb895ed00238