Right-align "end time" portion of the schedule setter.
This is more like what the mocks show. To make it happen, this changes the layout to a ConstraintLayout with a guideline in the middle to make sure that the start and end time sections take up no more than half the screen.

Confirmed that it wraps rather than overlapping when the text becomes too large.

Flag: android.app.modes_ui
Bug: 346396147
Test: manual
Change-Id: I56adbd24594121a9cb5b4374f2eeecdfd40f61e4