Don't truncate the title in Color Correction options
By default, the SelectorWithWidgetPreference hard coded the maxlines of
the title to be 2 lines. Updating the SelectorWithWidgetPreference to
accept changes on maxlines on title.

Set titleMaxLines as 2147483647 (Integer.MAX_VALUE) as there is no
alternative to clear the maxLines that are already set by the Preference

Bug: 356726764
Test: manually tested the color correction screen in Deutsch
Test: verify the Wifi Hotspot sceurity screen is not affected by this
change (still have 2 as the max lines of the title)
Flag: EXEMPT resource only

Change-Id: I1414b6212b75d3f28bef6aad50a31ea7861d5811