Settings: Fix the a11y focus issues in App Notifications subpage
There are duplicate a11y focuses in the App Notifications subpage.
See the bug for more details.

To fix the App, we will make the LayoutPreference unselectable,
which makes the internal FrameLayout unfocusable.

To fix the Switch, we can disable its focusability.
This will make the FrameLayout the only focusable view for the
Main Switch Bar. This a11y fix applies to all instances of
SettingsMainSwitchPreference in the Settings app.

Bug: 306725248
Flag: EXEMPT bugfix
Test:
atest SettingsMainSwitchPreferenceTest
Manually tested the UI. The SettingsMainSwitchPreference
is also used in `Display & Touch > Adaptive brightness`,
`Accessibility > Caption Preference`, etc. The toggle acts as
expected in the pages.

Change-Id: I52b4965ac0ffc8d1576e8bc650715eb80ab4e8c7