Touchpad: replace references to "trackpad" with "touchpad"
When we first added the touchpad settings we hadn't settled on whether
to refer to them as "touchpads" or "trackpads" yet. We've since settled
on "touchpad", but that's left us with some Settings code referring to
one and some to the other. Let's standardize on "touchpad" since that's
consistent with user-facing strings and the rest of the input codebase.

This CL doesn't change string identifiers, as I believe that would cause
them to get different Translation Console IDs and thus be retranslated.

Bug: 245989146
Test: m
Test: check strings and icons are still displayed correctly in the
      touchpad settings and tutorial, and that settings toggles and the
      tutorial still work
Test: atest SettingsRoboTests
Flag: EXEMPT refactor
Change-Id: I930e9fc85e09e03da3513e5f8443077fa8ea91b0